Moved majority of all opportunity testing routes from test_main to test_opportunity for better maintenance and organization
