+++ tests/test_main.py
@@ -45,24 +45,6 @@ def test_professor_profile(test_client: FlaskClient) -> None:
     assert "phone" in data
     assert "role" in data
 
-
-def test_professor_opportunity_cards(test_client: FlaskClient) -> None:
-
-    response = test_client.get("/getProfessorOpportunityCards/led")
-    assert response.status_code == 200
-
-    # Load the response data as JSON
-    data = json.loads(response.data.decode("utf-8"))
-
-    # Test that the "name" key exists
-    assert len(data.keys()) > 0
-    for eachCard in data["data"]:
-        assert "title" in eachCard
-        assert "body" in eachCard
-        assert "attributes" in eachCard
-        assert "id" in eachCard
-
-
 def test_get_opportunity(test_client: FlaskClient) -> None:
     response = test_client.get("/getOpportunity/2")
 
@@ -86,151 +68,7 @@ def test_get_opportunity(test_client: FlaskClient) -> None:
         assert "description" in eachSection
 
 
-def test_get_opportunity_meta(test_client: FlaskClient) -> None:
-    """
-    GIVEN a Flask application configured for testing
-    WHEN the '/getOpportunityMeta' endpoint is requested (GET) with valid data
-    THEN check that the response is valid and contains expected opportunity data
-    """
 
-    response = test_client.get("/getOpportunityMeta/1", content_type="application/json")
-
-    # assert response.status_code == 200
-
-    data = json.loads(response.data)
-    data = data["data"]
-
-    # Assertions on the expected data
-    assert "name" in data
-    assert "description" in data
-    assert "recommended_experience" in data
-    assert "pay" in data
-    assert "credits" in data
-    assert "semester" in data
-    assert "year" in data
-    assert "application_due" in data
-    assert "active" in data
-    assert "courses" in data
-    assert "majors" in data
-    assert "years" in data
-    assert "active" in data
-
-
-def test_get_opportunity_professor(test_client: FlaskClient) -> None:
-    response = test_client.get("/getOpportunityByProfessor/led")
-
-    assert response.status_code == 200
-
-    # Load the response data as JSON
-    data = json.loads(response.data.decode("utf-8"))
-    data = data["data"]
-
-    # Test that the "name" key exists
-    for opportunity in data:
-        assert "id" in opportunity
-        assert "name" in opportunity
-        assert "description" in opportunity
-        assert "recommended_experience" in opportunity
-        assert "pay" in opportunity
-        # assert "credits" in opportunity
-        assert "semester" in opportunity
-        assert "year" in opportunity
-        assert "application_due" in opportunity
-        assert "active" in opportunity
-        # assert "professor" in opportunity
-        # assert "department" in opportunity
-
-
-def test_get_professor_opportunity_cards(test_client: FlaskClient) -> None:
-    response = test_client.get(
-        "/getProfessorOpportunityCards/led", content_type="application/json"
-    )
-
-    assert response.status_code == 200
-
-    data = json.loads(response.data.decode("utf-8"))
-    data = data["data"]
-
-    for eachCard in data:
-        assert "title" in eachCard
-        assert "body" in eachCard
-        assert "attributes" in eachCard
-        assert "id" in eachCard
-
-
-def test_profile_opportunities(test_client: FlaskClient) -> None:
-    response = test_client.get(
-        "/getProfileOpportunities/led", content_type="application/json"
-    )
-
-    assert response.status_code == 200
-
-    data = json.loads(response.data.decode("utf-8"))
-    data = data["data"]
-
-    for eachCard in data:
-        assert "id" in eachCard
-        assert "title" in eachCard
-        assert "body" in eachCard
-        assert "attributes" in eachCard
-        assert "activeStatus" in eachCard
-
-
-def test_create_opportunity(test_client: FlaskClient) -> None:
-    """
-    GIVEN a Flask application configured for testing
-    WHEN the '/createOpportunity' endpoint is requested (POST) with valid data
-    THEN check that the response is valid and contains expected data
-    """
-
-    test_data = {
-        "authorID": "led",
-        "name": "Some test opportunity",
-        "description": "Some test description",
-        "recommended_experience": "Some test experience",
-        "pay": 25.0,
-        "credits": ["1", "2", "3", "4"],
-        "semester": "FALL",
-        "year": 2024,
-        "application_due": "2024-03-30",
-        "active": True,
-        "courses": ["CSCI4430"],
-        "majors": ["BIOL"],
-        "years": [2023, 2024],
-        "active": True,
-        "location": "TBD",
-    }
-
-    response = test_client.post(
-        "/createOpportunity",
-        data=json.dumps(test_data),
-        content_type="application/json",
-    )
-
-    assert response.status_code == 200
-
-    # query database to check for new opportunity with the same name
-    query = db.session.query(Opportunities).filter(
-        Opportunities.name == "Some test opportunity",
-        Opportunities.description == "Some test description",
-        Opportunities.recommended_experience == "Some test experience",
-    )
-
-    data = query.first()
-    id = data.id
-
-    # delete the opportunity by sending request to deleteOpportunity
-    response = test_client.post(
-        "/deleteOpportunity",
-        data=json.dumps({"id": id}),
-        content_type="application/json",
-    )
-
-    assert response.status_code == 200
-
-    # check that the opportunity was deleted
-    query = db.session.query(Opportunities).filter(Opportunities.id == id)
-    assert query.first() == None
 
 
 # def test_schools_and_departments(test_client: FlaskClient) -> None:
